+++ test/TemplatingTest.php
@@ -7,6 +7,7 @@ namespace Test;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
+use SubstancePHP\HTTP\EmptyShare;
 use SubstancePHP\HTTP\Templating;
 
 #[CoversClass(Templating::class)]
@@ -19,7 +20,7 @@ class TemplatingTest extends TestCase
         $this->assertSame('/templates', $templating->root);
         $this->assertSame('utf-8', $templating->encoding);
         $this->assertSame('layout', $templating->defaultLayout);
-        $this->assertSame([], $templating->shared);
+        $this->assertSame(EmptyShare::class, $templating->share);
         $this->assertNull($templating->assets);
     }
 }